Mehak commit

+++ Fincore.Infrastructure/Services/StateService.cs
@@ -1,4 +1,5 @@
-﻿using Fincore.Application.Interfaces;
+﻿using Fincore.Application.CommonHelper;
+using Fincore.Application.Interfaces;
 using Fincore.Domain.Models;
 using Fincore.Infrastructure.Data;
 using Microsoft.EntityFrameworkCore;
@@ -14,24 +15,41 @@ namespace Fincore.Infrastructure.Services
             _context = context;
         }
 
-        public async Task<IEnumerable<State>> GetAllAsync()
+        // Pagination
+        public async Task<PagedResponse<State>> GetAllAsync(int pageNumber, int pageSize)
         {
-            return await _context.States
-                .Include(s => s.Country)
+            var totalRecords = await _context.States.CountAsync();
+
+            var states = await _context.States
+                .Include(x => x.Country)
+                .OrderBy(x => x.StateId)
+                .Skip((pageNumber - 1) * pageSize)
+                .Take(pageSize)
                 .ToListAsync();
+
+            return new PagedResponse<State>
+            {
+                PageNumber = pageNumber,
+                PageSize = pageSize,
+                TotalRecords = totalRecords,
+                TotalPages = (int)Math.Ceiling((double)totalRecords / pageSize),
+                Data = states
+            };
         }
 
         public async Task<State?> GetByIdAsync(int id)
         {
             return await _context.States
-                .Include(s => s.Country)
-                .FirstOrDefaultAsync(s => s.StateId == id);
+                .Include(x => x.Country)
+                .FirstOrDefaultAsync(x => x.StateId == id);
         }
 
         public async Task<State> CreateAsync(State state)
         {
             _context.States.Add(state);
+
             await _context.SaveChangesAsync();
+
             return state;
         }
 
@@ -50,9 +68,6 @@ namespace Fincore.Infrastructure.Services
             return existingState;
         }
 
-
-
-
         public async Task<bool> DeleteAsync(int id)
         {
             var state = await _context.States.FindAsync(id);
@@ -60,13 +75,14 @@ namespace Fincore.Infrastructure.Services
             if (state == null)
                 return false;
 
-            var cities = await _context.Cities
-                .Where(c => c.StateId == id)
-                .ToListAsync();
+            // Validation
+            bool cityExists = await _context.Cities
+                .AnyAsync(x => x.StateId == id);
 
-            if (cities.Any())
+            if (cityExists)
             {
-                _context.Cities.RemoveRange(cities);
+                throw new InvalidOperationException(
+                    "Cannot delete this state because it contains cities.");
             }
 
             _context.States.Remove(state);
